evaluator: Move Runtime interface out of builtin.go
Move Runtime interface out of builtin.go into its own runtime.go,
including GraphicsRuntime interface and UnimplementedRuntime
implementation. This is in response to builtin.go getting rather huge.
